Extend column introspection to Cloud SQL Postgres and bust stale tree cache
